# frontend/openapi/core.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Accounts, deposits, transfers and reversals on a double-entry ledger, and the sign-in, user and API key management that guards them. Amounts are integers in the currency's minor unit. Errors are RFC 9457 problem details with a stable `code`.",
    "title" : "PayLedger Core API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/accounts" : {
      "get" : {
        "description" : "**Access:** `hasAnyRole('CUSTOMER', 'OPERATOR')`",
        "operationId" : "account_list",
        "parameters" : [ {
          "in" : "query",
          "name" : "ownerId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AccountResponse"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      },
      "post" : {
        "description" : "**Access:** `hasRole('CUSTOMER')`",
        "operationId" : "account_open",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OpenAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "Opened"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/accounts/{id}" : {
      "get" : {
        "description" : "**Access:** `hasAnyRole('CUSTOMER', 'OPERATOR')`",
        "operationId" : "account_get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/accounts/{id}/close" : {
      "post" : {
        "description" : "**Access:** `hasAnyRole('CUSTOMER', 'OPERATOR')`",
        "operationId" : "account_close",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/accounts/{id}/freeze" : {
      "post" : {
        "description" : "**Access:** `hasRole('OPERATOR')`",
        "operationId" : "account_freeze",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/accounts/{id}/ledger-entries" : {
      "get" : {
        "description" : "**Access:** `hasAnyRole('CUSTOMER', 'OPERATOR')`",
        "operationId" : "accountStatement_ledgerEntries",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "startingAfter",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 100,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StatementPage"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/accounts/{id}/unfreeze" : {
      "post" : {
        "description" : "**Access:** `hasRole('OPERATOR')`",
        "operationId" : "account_unfreeze",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "Accounts" ]
      }
    },
    "/api/v1/api-keys" : {
      "get" : {
        "description" : "**Access:** `hasRole('ADMIN')`",
        "operationId" : "apiKey_list",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ApiKey"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "API keys" ]
      },
      "post" : {
        "description" : "**Access:** `hasRole('ADMIN')`",
        "operationId" : "apiKey_create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedApiKeyResponse"
                }
              }
            },
            "description" : "Issued"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "API keys" ]
      }
    },
    "/api/v1/api-keys/{id}" : {
      "get" : {
        "description" : "**Access:** `hasRole('ADMIN')`",
        "operationId" : "apiKey_get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiKey"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "API keys" ]
      }
    },
    "/api/v1/api-keys/{id}/revoke" : {
      "post" : {
        "description" : "**Access:** `hasRole('ADMIN')`",
        "operationId" : "apiKey_revoke",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiKey"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "An RFC 9457 problem; `code` says what went wrong"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "API keys" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AccountResponse" : {
        "type" : "object",
        "properties" : {
          "balance" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "currency" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ownerId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "FROZEN", "CLOSED" ]
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CUSTOMER", "SYSTEM" ]
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "balance", "createdAt", "currency", "id", "ownerId", "status", "type", "updatedAt" ]
      },
      "ApiKey" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "lastUsedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "name" : {
            "type" : "string"
          },
          "prefix" : {
            "type" : "string"
          },
          "revokedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "deposits:write" ]
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "createdAt", "createdBy", "expiresAt", "id", "lastUsedAt", "name", "prefix", "revokedAt", "scopes" ]
      },
...
      "CreateApiKeyRequest" : {
        "type" : "object",
        "properties" : {
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "deposits:write" ]
            },
            "minItems" : 1,
            "uniqueItems" : true
          }
        },
        "required" : [ "name", "scopes" ]
      },
...
      "CreatedApiKeyResponse" : {
        "type" : "object",
        "properties" : {
          "apiKey" : {
            "$ref" : "#/components/schemas/ApiKey"
          },
          "key" : {
            "type" : "string"
          }
        },
        "required" : [ "apiKey", "key" ]
      },
...
      "InvalidParam" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        },
        "required" : [ "name", "reason" ]
      },
...
      "OpenAccountRequest" : {
        "type" : "object",
        "properties" : {
          "currency" : {
            "type" : "string",
            "minLength" : 1,
            "pattern" : "^[A-Z]{3}$"
          }
        },
        "required" : [ "currency" ]
      },
      "Problem" : {
        "type" : "object",
        "description" : "RFC 9457 problem details with a stable `code`",
        "properties" : {
          "code" : {
            "type" : "string",
            "example" : "INSUFFICIENT_FUNDS"
          },
          "detail" : {
            "type" : "string",
            "description" : "For a person, not for code to parse"
          },
          "instance" : {
            "type" : "string",
            "description" : "The request path"
          },
          "invalidParams" : {
            "type" : "array",
            "description" : "The fields that failed validation (code INVALID_REQUEST), as in RFC 9457's example",
            "items" : {
              "$ref" : "#/components/schemas/InvalidParam"
            }
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 422
          },
          "title" : {
            "type" : "string",
            "example" : "Unprocessable Content"
          },
          "transferId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "The FAILED movement recorded when a business rule rejected a deposit, transfer or reversal"
          },
          "type" : {
            "type" : "string",
            "example" : "about:blank"
          }
        },
        "required" : [ "code", "status", "title", "type" ]
      },
...
      "StatementEntry" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "balanceAfter" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "The account's balance right after this entry"
          },
          "counterpartyAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "direction" : {
            "type" : "string",
            "description" : "DEBIT takes money out of the account, CREDIT puts it in",
            "enum" : [ "DEBIT", "CREDIT" ]
          },
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Entry id; pass the last one as startingAfter for the next page"
          },
          "transferId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "transferStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "REVERSED" ]
          },
          "type" : {
            "type" : "string",
            "enum" : [ "DEPOSIT", "TRANSFER", "REVERSAL" ]
          }
        },
        "required" : [ "amount", "balanceAfter", "counterpartyAccountId", "createdAt", "currency", "description", "direction", "id", "transferId", "transferStatus", "type" ]
      },
      "StatementPage" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StatementEntry"
            }
          },
          "hasMore" : {
            "type" : "boolean"
          }
        },
        "required" : [ "data", "hasMore" ]
      },
...
    },
    "securitySchemes" : {
      "apiKey" : {
        "description" : "A machine client's key, e.g. the partner bank's, with its scopes",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "An ES256 access token (RFC 9068) from /api/v1/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}